style: added images to About Me section and also move folders arounf to suit next.js file structure
