+++ src/app/aiassistant/chatbox.tsx
