+++ src/app/aiassistant/page.tsx
@@ -1,13 +1,14 @@
 import Header from "../components/header/header";
 import Footer from "../components/footer/footer";
 import Navigation from "../components/navigation/navigation";
+import { Chatbox } from "@/app/aiassistant/chatbox";
 
-export default function Tree() {
+export default function Assistant() {
   return (
     <div>
       <Header></Header>
       <Navigation></Navigation>
-      <h1>Tree</h1>
+      <Chatbox />
       <Footer></Footer>
     </div>
   );
