+++ src/app/calendar/page.tsx
@@ -1,6 +1,6 @@
-import Header from "../../components/header/header";
-import Footer from "../../components/footer/footer";
-import Navigation from "../../components/navigation/navigation";
+import Header from "../components/header/header";
+import Footer from "../components/footer/footer";
+import Navigation from "../components/navigation/navigation";
 import CalendarComponent from "../../app/calendar/CalendarComponent";
 
 export default function Calendar() {
