+++ src/app/components/DatabaseInitializer.tsx
