+++ src/app/components/auth/LogOutButton.tsx
