+++ src/app/components/auth/SignInForm.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react";
 import { useRouter } from "next/navigation";
 import Link from "next/link";
-import { supabase } from "../../lib/supabaseClient";
+import { supabase } from "../../../lib/supabaseClient";
 import styles from "./SignInForm.module.css";
 import { GoogleSignIn } from "./GoogleSignIn";
