+++ src/app/components/navigation/navigation.tsx
@@ -1,33 +1,3 @@
-// import Link from "next/link";
-// import styles from "./navigation.module.css";
-
-// export default function Navigation() {
-//   return (
-//     <nav>
-//       <ul className={styles.navList}>
-//         <li>
-//           <Link href="/">Home</Link>
-//         </li>
-//         <li>
-//           <Link href="/timeline">Timeline</Link>
-//         </li>
-//         <li>
-//           <Link href="/about">About</Link>
-//         </li>
-//         <li>
-//           <Link href="/calendar">Calendar</Link>
-//         </li>
-//         <li>
-//           <Link href="/tree">Tree</Link>
-//         </li>
-//         <li>
-//           <Link href="/assistant">Assistant</Link>
-//         </li>
-//       </ul>
-//     </nav>
-//   );
-// }
-
 import Link from "next/link";
 import styles from "./navigation.module.css";
 import LogoutButton from "../auth/LogOutButton";
@@ -63,7 +33,7 @@ export default function Navigation() {
             </Link>
           </li>
           <li className={styles.navItem}>
-            <Link href="/assistant" className={styles.navLink}>
+            <Link href="/aiassistant" className={styles.navLink}>
               Assistant
             </Link>
           </li>
